+++ src/main/php/lang/ast/syntax/php/ThrowExpressions.class.php
@@ -19,7 +19,7 @@ use lang\ast\syntax\Extension;
 class ThrowExpressions implements Extension {
 
   public function setup($language, $emitter) {
-    $language->prefix('throw', function($parse, $token) {
+    $language->prefix('throw', 0, function($parse, $token) {
       return new ThrowExpression($this->expression($parse, 0), $token->line);
     });
   }